import z/osmf cert; run post verify after zowe is started

+++ scripts/temp-fixes-after-install.sh
@@ -36,22 +36,31 @@ echo "[${SCRIPT_NAME}] change ${CI_ZOWE_DS_MEMBER} RACF user ..."
 (exec sh -c "tsocmd \"SETROPTS RACLIST(STARTED) REFRESH\"")
 echo
 
+echo
+echo "[${SCRIPT_NAME}] import z/OSMF certificates which requires superuser permission ..."
+(exec sh -c "cd ${CI_ZOWE_ROOT_DIR}/api-mediation && su && scripts/apiml_cm.sh --action trust-zosmf --zosmf-keyring IZUKeyring.IZUDFLT --zosmf-userid IZUSVR")
+echo
+
 ################################################################################
 # explorer JES/MVS/USS has internal host name, convert to public domain
 echo
 echo "[${SCRIPT_NAME}] checking hostname in explorer-* ..."
 ZDNT_HOSTNAME=S0W1
-FILES_TO_UPDATE="explorer-JES explorer-USS explorer-MVS api_catalog"
+FILES_TO_UPDATE="explorer-JES explorer-USS explorer-MVS api_catalog jes_explorer"
 for one in $FILES_TO_UPDATE; do
   ZDNT_FILE=$CI_ZOWE_ROOT_DIR/${one}/web/index.html
   echo "[${SCRIPT_NAME}]   - checking $ZDNT_FILE ..."
-  HAS_WRONG_HOSTNAME=$(grep $ZDNT_HOSTNAME $ZDNT_FILE)
-  if [ -n "$HAS_WRONG_HOSTNAME" ]; then
-    sed "s#//${ZDNT_HOSTNAME}:\([0-9]\+\)/#//${CI_HOSTNAME}:\1/#" $ZDNT_FILE > index.html.tmp
-    mv index.html.tmp $ZDNT_FILE
-    echo "[${SCRIPT_NAME}]     - updated."
+  if [ -f "$ZDNT_FILE" ]; then
+    HAS_WRONG_HOSTNAME=$(grep $ZDNT_HOSTNAME $ZDNT_FILE)
+    if [ -n "$HAS_WRONG_HOSTNAME" ]; then
+      sed "s#//${ZDNT_HOSTNAME}:\([0-9]\+\)/#//${CI_HOSTNAME}:\1/#" $ZDNT_FILE > index.html.tmp
+      mv index.html.tmp $ZDNT_FILE
+      echo "[${SCRIPT_NAME}]     - updated."
+    else
+      echo "[${SCRIPT_NAME}]     - no need to update."
+    fi
   else
-    echo "[${SCRIPT_NAME}]     - no need to update."
+    echo "[${SCRIPT_NAME}]     - doesn't exist."
   fi
 done
